Started the AI Class
We're going to have to restructure how the Legal Vectors work.  I think
I'm going to bring in the Location class we used in other projects,
rather than using a one dimensional array.  Also, we need to add in
checks to ensure that the vectors are within the board, otherwise we get
null pointer exceptions.