Got an AI working
I have an AI working only for white, but it does work.  It shouldn't be
too hard to get it working for both white and black.  It only reads it's
own moves and the opponents possible responses, and the position
analysis is very simplistic.  And it's slow.  Even on my computer it
takes a few seconds for each move, I don't know how bad it will be on an
emulator.